ChatGPT prompting functions
Defined functions in the ChatGPT_prompting.py script that creates a response from ChatGPT from a prompting message, a context windows (a list of message structured variables) and a chosen role for the message.

One of the functions is designed for being called in terms of single user interactment with chatgpt, while the other function (Simple_PromptChatGPT_ReturnResponse) is simpler and can be used for testing or when knowing who talked to ChatGPT is not important.