Head Orientations for discussion round +Orient Head ID translations
Added code to the main script which saves head orientations into variables that can easily be used for the discussion round (since we know the order of participants in discussion round). For this I had to add a translation method in the ChangeHeadOrientation.py script, so that participation IDs can be translated to head orientations for the robot.